Move C++ helper header to utils
