+++ scripts/qemu/configs/default.xml
@@ -50,16 +50,12 @@ Cell 2 |   0   |   1200MiB    | 17, 28, 10          |
       <interconnects>
         <latency initiator='0' target='0' type='access' value='10'/>
         <bandwidth initiator='0' target='0' type='access' value='10485760' unit='B'/>
-        
         <latency initiator='0' target='1' type='access' value='20'/>
         <bandwidth initiator='0' target='1' type='access' value='5242880' unit='B'/>
-        
         <latency initiator='0' target='2' type='access' value='16'/>
         <bandwidth initiator='0' target='2' type='access' value='1048576' unit='B'/>
-       
         <latency initiator='1' target='1' type='access' value='10'/>
         <bandwidth initiator='1' target='1' type='access' value='10485760' unit='B'/>
-        
         <latency initiator='1' target='2' type='access' value='27'/>
         <bandwidth initiator='1' target='2' type='access' value='1048576' unit='B'/>
       </interconnects>
