+++ test/provider_fixed_memory.cpp
@@ -4,8 +4,8 @@
 
 #include "base.hpp"
 
-#include "cpp_helpers.hpp"
 #include "test_helpers.h"
+#include "utils/cpp_helpers.hpp"
 #ifndef _WIN32
 #include "test_helpers_linux.h"
 #endif
