+++ test/provider_os_memory.cpp
@@ -4,9 +4,9 @@
 
 #include "base.hpp"
 
-#include "cpp_helpers.hpp"
 #include "ipcFixtures.hpp"
 #include "test_helpers.h"
+#include "utils/cpp_helpers.hpp"
 
 #include <umf/memory_provider.h>
 #include <umf/pools/pool_disjoint.h>
